4-Shopping_Agent/pyproject.toml
[project]
name = "4-shopping-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.7.1.1",
    "openai-agents>=0.2.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
